ai-agent/requirement.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pymongo==4.12.0
langchain==0.3.23
chromadb==1.0.5
google-genai==1.11.0
python-dotenv==1.1.0
